+++ tests/goal/test_goal.py
@@ -3,7 +3,8 @@ from pathlib import Path
 from subprocess import CompletedProcess
 
 import pytest
-from algokit.core.sandbox import get_algod_network_template, get_config_json, get_docker_compose_yml
+from algokit.core.sandbox import ALGOD_HEALTH_URL, get_algod_network_template, get_config_json, get_docker_compose_yml
+from pytest_httpx import HTTPXMock
 from pytest_mock import MockerFixture
 
 from tests.utils.app_dir_mock import AppDirs
@@ -19,6 +20,11 @@ def _normalize_output(output: str) -> str:
 
 
 @pytest.fixture()
+def _health_success(httpx_mock: HTTPXMock) -> None:
+    httpx_mock.add_response(url=ALGOD_HEALTH_URL)
+
+
+@pytest.fixture()
 def cwd(tmp_path_factory: pytest.TempPathFactory) -> Path:
     return tmp_path_factory.mktemp("cwd")
 
@@ -27,7 +33,7 @@ def cwd(tmp_path_factory: pytest.TempPathFactory) -> Path:
 def mocked_goal_mount_path(cwd: Path, monkeypatch: pytest.MonkeyPatch) -> Path:
     mocked_goal_mount = cwd / "goal_mount"
     mocked_goal_mount.mkdir()
-    monkeypatch.setattr("algokit.cli.goal.get_volume_mount_path_local", lambda: cwd / "goal_mount")
+    monkeypatch.setattr("algokit.cli.goal.get_volume_mount_path_local", lambda directory_name: cwd / "goal_mount")  # noqa: ARG005
     return mocked_goal_mount
 
 
@@ -50,6 +56,19 @@ def _setup_input_files(cwd: Path, request: pytest.FixtureRequest) -> None:
                 (cwd / file["name"]).touch()
 
             assert (cwd / file["name"]).exists()
+
+
+@pytest.fixture()
+def _mock_proc_with_running_localnet(proc_mock: ProcMock) -> None:
+    proc_mock.set_output("docker compose ls --format json --filter name=algokit_sandbox*", [json.dumps([])])
+
+
+@pytest.fixture()
+def _mock_proc_with_algod_running_state(proc_mock: ProcMock) -> None:
+    proc_mock.set_output(
+        cmd=["docker", "compose", "ps", "algod", "--format", "json"],
+        output=[json.dumps([{"Name": "algokit_sandbox_algod", "State": "running"}])],
+    )
 
 
 def dump_file(cwd: Path) -> None:
@@ -77,16 +96,27 @@ def test_goal_help() -> None:
     verify(result.output)
 
 
-@pytest.mark.usefixtures("proc_mock", "_setup_latest_dummy_compose", "mocked_goal_mount_path")
-def test_goal_no_args() -> None:
+@pytest.mark.usefixtures(
+    "proc_mock",
+    "_setup_latest_dummy_compose",
+    "mocked_goal_mount_path",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
+def test_goal_no_args(app_dir_mock: AppDirs) -> None:
     result = invoke("goal")
 
     assert result.exit_code == 0
-    verify(result.output)
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
 
 
-@pytest.mark.usefixtures("proc_mock", "_setup_latest_dummy_compose")
-def test_goal_console(mocker: MockerFixture) -> None:
+@pytest.mark.usefixtures(
+    "proc_mock",
+    "_setup_latest_dummy_compose",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
+def test_goal_console(mocker: MockerFixture, app_dir_mock: AppDirs) -> None:
     mocker.patch("algokit.core.proc.subprocess_run").return_value = CompletedProcess(
         ["docker", "exec"], 0, "STDOUT+STDERR"
     )
@@ -94,56 +124,66 @@ def test_goal_console(mocker: MockerFixture) -> None:
     result = invoke("goal --console")
 
     assert result.exit_code == 0
-    verify(result.output)
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
 
 
-@pytest.mark.usefixtures("_setup_latest_dummy_compose")
-def test_goal_console_failed(app_dir_mock: AppDirs, proc_mock: ProcMock, mocker: MockerFixture) -> None:
+@pytest.mark.usefixtures("_setup_latest_dummy_compose", "_mock_proc_with_running_localnet", "_health_success")
+def test_goal_console_algod_not_created(app_dir_mock: AppDirs, proc_mock: ProcMock, mocker: MockerFixture) -> None:
+    proc_mock.set_output(["docker", "compose", "ps", "algod", "--format", "json"], output=[json.dumps([])])
+
+    mocker.patch("algokit.core.proc.subprocess_run").return_value = CompletedProcess(
+        ["docker", "exec"], 0, "STDOUT+STDERR"
+    )
+
+    result = invoke("goal --console")
+
+    assert result.exit_code == 0
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
+
+
+@pytest.mark.usefixtures(
+    "proc_mock",
+    "_setup_latest_dummy_compose",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
+def test_goal_console_failed(app_dir_mock: AppDirs, mocker: MockerFixture) -> None:
     mocker.patch("algokit.core.proc.subprocess_run").return_value = CompletedProcess(
         ["docker", "exec"], 1, "STDOUT+STDERR"
     )
 
-    proc_mock.set_output(
-        ["docker", "compose", "ps", "algod", "--format", "json"],
-        output=[json.dumps([{"Name": "algokit_algod", "State": "running"}])],
-    )
-
     result = invoke("goal --console")
 
     assert result.exit_code == 1
     verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
 
 
-@pytest.mark.usefixtures("_setup_latest_dummy_compose")
-def test_goal_console_failed_algod_not_created(
-    app_dir_mock: AppDirs, proc_mock: ProcMock, mocker: MockerFixture
-) -> None:
-    mocker.patch("algokit.core.proc.subprocess_run").return_value = CompletedProcess(
-        ["docker", "exec"], 1, "bad args to goal"
-    )
-
-    proc_mock.set_output(["docker", "compose", "ps", "algod", "--format", "json"], output=[json.dumps([])])
-
-    result = invoke("goal --console")
-
-    assert result.exit_code == 1
-    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
-
-
-@pytest.mark.usefixtures("proc_mock", "_setup_latest_dummy_compose", "mocked_goal_mount_path")
-def test_goal_simple_args() -> None:
+@pytest.mark.usefixtures(
+    "proc_mock",
+    "_setup_latest_dummy_compose",
+    "mocked_goal_mount_path",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
+def test_goal_simple_args(app_dir_mock: AppDirs) -> None:
     result = invoke("goal account list")
 
     assert result.exit_code == 0
-    verify(result.output)
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
 
 
-@pytest.mark.usefixtures("proc_mock", "_setup_latest_dummy_compose", "mocked_goal_mount_path")
-def test_goal_complex_args() -> None:
+@pytest.mark.usefixtures(
+    "proc_mock",
+    "_setup_latest_dummy_compose",
+    "mocked_goal_mount_path",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
+def test_goal_complex_args(app_dir_mock: AppDirs) -> None:
     result = invoke("goal account export -a RKTAZY2ZLKUJBHDVVA3KKHEDK7PRVGIGOZAUUIZBNK2OEP6KQGEXKKUYUY")
 
     assert result.exit_code == 0
-    verify(result.output)
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
 
 
 def test_goal_start_without_docker(proc_mock: ProcMock) -> None:
@@ -164,11 +204,18 @@ def test_goal_start_without_docker_engine_running(proc_mock: ProcMock) -> None:
     verify(result.output)
 
 
-@pytest.mark.usefixtures("_setup_input_files", "_setup_latest_dummy_compose", "mocked_goal_mount_path")
+@pytest.mark.usefixtures(
+    "_setup_input_files",
+    "_setup_latest_dummy_compose",
+    "mocked_goal_mount_path",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
 @pytest.mark.parametrize("_setup_input_files", [[{"name": "transactions.txt"}]], indirect=True)
 def test_goal_simple_args_with_input_file(
     proc_mock: ProcMock,
     cwd: Path,
+    app_dir_mock: AppDirs,
 ) -> None:
     expected_arguments = [
         "docker",
@@ -176,7 +223,7 @@ def test_goal_simple_args_with_input_file(
         "--interactive",
         "--workdir",
         "/root",
-        "algokit_algod",
+        "algokit_sandbox_algod",
         "goal",
         "clerk",
         "group",
@@ -186,23 +233,28 @@ def test_goal_simple_args_with_input_file(
     result = invoke("goal clerk group transactions.txt", cwd=cwd)
 
     # Check if the path in command has changed in preprocess step
-    assert _normalize_output(proc_mock.called[1].command[9]) == "/root/goal_mount/transactions.txt"
+    assert _normalize_output(proc_mock.called[3].command[9]) == "/root/goal_mount/transactions.txt"
 
     # Check for the result status
     assert result.exit_code == 0
 
-    verify(_normalize_output(result.output))
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
 
 
-@pytest.mark.usefixtures("mocked_goal_mount_path", "_setup_latest_dummy_compose")
-def test_goal_simple_args_with_output_file(proc_mock: ProcMock, cwd: Path) -> None:
+@pytest.mark.usefixtures(
+    "mocked_goal_mount_path",
+    "_setup_latest_dummy_compose",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
+def test_goal_simple_args_with_output_file(proc_mock: ProcMock, cwd: Path, app_dir_mock: AppDirs) -> None:
     expected_arguments = [
         "docker",
         "exec",
         "--interactive",
         "--workdir",
         "/root",
-        "algokit_algod",
+        "algokit_sandbox_algod",
         "goal",
         "account",
         "dump",
@@ -217,7 +269,7 @@ def test_goal_simple_args_with_output_file(proc_mock: ProcMock, cwd: Path) -> No
     result = invoke("goal account dump -o balance_record.json")
 
     # Check if the path in command has changed in preprocess step
-    assert _normalize_output(proc_mock.called[1].command[10]) == "/root/goal_mount/balance_record.json"
+    assert _normalize_output(proc_mock.called[3].command[10]) == "/root/goal_mount/balance_record.json"
 
     # Check for the result status
     assert result.exit_code == 0
@@ -225,16 +277,23 @@ def test_goal_simple_args_with_output_file(proc_mock: ProcMock, cwd: Path) -> No
     # Check if the output file is actually created and copied in cwd in postprocess step
     assert (cwd / "balance_record.json").exists()
 
-    verify(_normalize_output(result.output))
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
 
 
-@pytest.mark.usefixtures("mocked_goal_mount_path", "_setup_input_files", "_setup_latest_dummy_compose")
+@pytest.mark.usefixtures(
+    "mocked_goal_mount_path",
+    "_setup_input_files",
+    "_setup_latest_dummy_compose",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
 @pytest.mark.parametrize(
     "_setup_input_files", [[{"name": "approval.teal", "content": DUMMY_CONTRACT_TEAL}]], indirect=True
 )
 def test_goal_simple_args_with_input_output_files(
     proc_mock: ProcMock,
     cwd: Path,
+    app_dir_mock: AppDirs,
 ) -> None:
     expected_arguments = [
         "docker",
@@ -242,7 +301,7 @@ def test_goal_simple_args_with_input_output_files(
         "--interactive",
         "--workdir",
         "/root",
-        "algokit_algod",
+        "algokit_sandbox_algod",
         "goal",
         "clerk",
         "compile",
@@ -255,18 +314,24 @@ def test_goal_simple_args_with_input_output_files(
     result = invoke("goal clerk compile approval.teal -o approval.compiled", cwd=cwd)
 
     # Check if the paths in command have changed in preprocess step
-    assert _normalize_output(proc_mock.called[1].command[9]) == "/root/goal_mount/approval.teal"
-    assert _normalize_output(proc_mock.called[1].command[11]) == "/root/goal_mount/approval.compiled"
+    assert _normalize_output(proc_mock.called[3].command[9]) == "/root/goal_mount/approval.teal"
+    assert _normalize_output(proc_mock.called[3].command[11]) == "/root/goal_mount/approval.compiled"
 
     # Check for the result status
     assert result.exit_code == 0
 
     # Check if the output file is created and copied in cwd in postprocess step
     assert (cwd / "approval.compiled").exists()
-    verify(_normalize_output(result.output))
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
 
 
-@pytest.mark.usefixtures("mocked_goal_mount_path", "_setup_input_files", "_setup_latest_dummy_compose")
+@pytest.mark.usefixtures(
+    "mocked_goal_mount_path",
+    "_setup_input_files",
+    "_setup_latest_dummy_compose",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
 @pytest.mark.parametrize(
     "_setup_input_files",
     [
@@ -280,6 +345,7 @@ def test_goal_simple_args_with_input_output_files(
 def test_goal_simple_args_with_multiple_input_output_files(
     proc_mock: ProcMock,
     cwd: Path,
+    app_dir_mock: AppDirs,
 ) -> None:
     expected_arguments = [
         "docker",
@@ -287,7 +353,7 @@ def test_goal_simple_args_with_multiple_input_output_files(
         "--interactive",
         "--workdir",
         "/root",
-        "algokit_algod",
+        "algokit_sandbox_algod",
         "goal",
         "clerk",
         "compile",
@@ -299,30 +365,42 @@ def test_goal_simple_args_with_multiple_input_output_files(
     result = invoke("goal clerk compile approval1.teal approval2.teal -o approval.compiled", cwd=cwd)
 
     # Check if the paths in command have changed in preprocess step
-    assert _normalize_output(proc_mock.called[1].command[9]) == "/root/goal_mount/approval1.teal"
-    assert _normalize_output(proc_mock.called[1].command[10]) == "/root/goal_mount/approval2.teal"
-    assert _normalize_output(proc_mock.called[1].command[12]) == "/root/goal_mount/approval.compiled"
+    assert _normalize_output(proc_mock.called[3].command[9]) == "/root/goal_mount/approval1.teal"
+    assert _normalize_output(proc_mock.called[3].command[10]) == "/root/goal_mount/approval2.teal"
+    assert _normalize_output(proc_mock.called[3].command[12]) == "/root/goal_mount/approval.compiled"
 
     # Check for the result
     assert result.exit_code == 0
 
     # Check if the output file is actually created and copied in cwd in postprocess step
     assert (cwd / "approval.compiled").exists()
-    verify(_normalize_output(result.output))
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
 
 
-@pytest.mark.usefixtures("proc_mock", "mocked_goal_mount_path", "_setup_latest_dummy_compose")
+@pytest.mark.usefixtures(
+    "proc_mock",
+    "mocked_goal_mount_path",
+    "_setup_latest_dummy_compose",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
 def test_goal_simple_args_without_file_error(
     cwd: Path,
+    app_dir_mock: AppDirs,
 ) -> None:
     assert not (cwd / "approval.teal").exists()
     result = invoke("goal clerk compile approval.teal -o approval.compiled", cwd=cwd)
 
     assert result.exit_code == 1
-    verify(_normalize_output(result.output))
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))
 
 
-@pytest.mark.usefixtures("_setup_input_files", "_setup_latest_dummy_compose")
+@pytest.mark.usefixtures(
+    "_setup_input_files",
+    "_setup_latest_dummy_compose",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
 @pytest.mark.parametrize(
     "_setup_input_files", [[{"name": "approval.teal", "content": DUMMY_CONTRACT_TEAL}]], indirect=True
 )
@@ -342,7 +420,7 @@ def test_goal_postprocess_of_command_args(
         "--interactive",
         "--workdir",
         "/root",
-        "algokit_algod",
+        "algokit_sandbox_algod",
         "goal",
         "clerk",
         "compile",
@@ -370,7 +448,12 @@ def test_goal_postprocess_of_command_args(
     assert (mocked_goal_mount_path / "approval.group.sig.out").exists()
 
 
-@pytest.mark.usefixtures("_setup_input_files", "_setup_latest_dummy_compose")
+@pytest.mark.usefixtures(
+    "_setup_input_files",
+    "_setup_latest_dummy_compose",
+    "_mock_proc_with_running_localnet",
+    "_mock_proc_with_algod_running_state",
+)
 @pytest.mark.parametrize("_setup_input_files", [[{"name": "group.gtxn", "content": ""}]], indirect=True)
 def test_goal_postprocess_of_single_output_arg_resulting_in_multiple_output_files(
     proc_mock: ProcMock,
@@ -383,7 +466,7 @@ def test_goal_postprocess_of_single_output_arg_resulting_in_multiple_output_file
         "--interactive",
         "--workdir",
         "/root",
-        "algokit_algod",
+        "algokit_sandbox_algod",
         "goal",
         "clerk",
         "split",
@@ -413,7 +496,7 @@ def test_goal_postprocess_of_single_output_arg_resulting_in_multiple_output_file
     assert (cwd / "group-1.txn").exists()
 
 
-@pytest.mark.usefixtures("proc_mock")
+@pytest.mark.usefixtures("proc_mock", "_mock_proc_with_running_localnet")
 def test_goal_compose_outdated(
     cwd: Path,
     app_dir_mock: AppDirs,
@@ -427,3 +510,16 @@ def test_goal_compose_outdated(
     assert result.exit_code == 1
 
     verify(_normalize_output(result.output))
+
+
+@pytest.mark.usefixtures("_setup_latest_dummy_compose", "mocked_goal_mount_path", "_mock_proc_with_algod_running_state")
+def test_goal_simple_args_on_named_localnet(proc_mock: ProcMock, app_dir_mock: AppDirs) -> None:
+    proc_mock.set_output(
+        "docker compose ls --format json --filter name=algokit_sandbox*",
+        [json.dumps([{"Name": "algokit_test", "Status": "running", "ConfigFiles": "to/test/docker-compose.yml"}])],
+    )
+
+    result = invoke("goal account list")
+
+    assert result.exit_code == 0
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))