+++ tests/localnet/test_localnet_console.py
@@ -1,18 +1,20 @@
+import json
 from subprocess import CompletedProcess
 
 import pytest
 from algokit.core.sandbox import get_algod_network_template, get_config_json, get_docker_compose_yml
 from pytest_mock import MockerFixture
 
+from tests.goal.test_goal import _normalize_output
 from tests.utils.app_dir_mock import AppDirs
 from tests.utils.approvals import verify
 from tests.utils.click_invoker import invoke
+from tests.utils.proc_mock import ProcMock
 
 
-@pytest.mark.usefixtures(
-    "proc_mock",
-)
-def test_goal_console(mocker: MockerFixture, tmp_path_factory: pytest.TempPathFactory, app_dir_mock: AppDirs) -> None:
+def test_goal_console(
+    mocker: MockerFixture, tmp_path_factory: pytest.TempPathFactory, app_dir_mock: AppDirs, proc_mock: ProcMock
+) -> None:
     cwd = tmp_path_factory.mktemp("cwd")
 
     mocked_goal_mount = cwd / "goal_mount"
@@ -27,8 +29,26 @@ def test_goal_console(mocker: MockerFixture, tmp_path_factory: pytest.TempPathFa
     mocker.patch("algokit.core.proc.subprocess_run").return_value = CompletedProcess(
         ["docker", "exec"], 0, "STDOUT+STDERR"
     )
+    proc_mock.set_output(
+        "docker compose ls --format json --filter name=algokit_sandbox*",
+        [
+            json.dumps(
+                [
+                    {
+                        "Name": "algokit_sandbox",
+                        "Status": "running",
+                        "ConfigFiles": "test/sandbox_test/docker-compose.yml",
+                    }
+                ]
+            )
+        ],
+    )
+    proc_mock.set_output(
+        cmd=["docker", "compose", "ps", "algod", "--format", "json"],
+        output=[json.dumps([{"Name": "algokit_sandbox_algod", "State": "running"}])],
+    )
 
     result = invoke("localnet console")
 
     assert result.exit_code == 0
-    verify(result.output)
+    verify(_normalize_output(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}")))