+++ tests/localnet/test_localnet_start.py
@@ -54,7 +54,7 @@ def _localnet_img_check_cmd_error(proc_mock: ProcMock) -> None:
     proc_mock.should_fail_on(["docker", "image", "inspect", INDEXER_IMAGE, "--format", arg])
 
 
-@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_up_to_date")
+@pytest.mark.usefixtures("_health_success", "_localnet_up_to_date", "_mock_proc_with_running_localnet")
 def test_localnet_start(app_dir_mock: AppDirs) -> None:
     result = invoke("localnet start")
 
@@ -68,7 +68,35 @@ def test_localnet_start(app_dir_mock: AppDirs) -> None:
     )
 
 
-@pytest.mark.usefixtures("proc_mock", "_localnet_up_to_date")
+@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_up_to_date")
+def test_localnet_start_with_name(app_dir_mock: AppDirs, proc_mock: ProcMock) -> None:
+    proc_mock.set_output(
+        "docker compose ls --format json --filter name=algokit_sandbox*",
+        [
+            json.dumps(
+                [
+                    {
+                        "Name": "algokit_sandbox_test",
+                        "Status": "running",
+                        "ConfigFiles": "sandbox_test/docker-compose.yml",
+                    }
+                ]
+            )
+        ],
+    )
+    result = invoke("localnet start --name test")
+
+    assert result.exit_code == 0
+    verify(
+        get_combined_verify_output(
+            result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}").replace("\\", "/"),
+            "{app_config}/sandbox_test/docker-compose.yml",
+            (app_dir_mock.app_config_dir / "sandbox_test" / "docker-compose.yml").read_text(),
+        )
+    )
+
+
+@pytest.mark.usefixtures("proc_mock", "_localnet_up_to_date", "_mock_proc_with_running_localnet")
 def test_localnet_start_health_failure(app_dir_mock: AppDirs, httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_exception(httpx.RemoteProtocolError("No response"), url=ALGOD_HEALTH_URL)
     result = invoke("localnet start")
@@ -83,7 +111,7 @@ def test_localnet_start_health_failure(app_dir_mock: AppDirs, httpx_mock: HTTPXM
     )
 
 
-@pytest.mark.usefixtures("proc_mock", "_localnet_up_to_date")
+@pytest.mark.usefixtures("proc_mock", "_localnet_up_to_date", "_mock_proc_with_running_localnet")
 def test_localnet_start_health_bad_status(app_dir_mock: AppDirs, httpx_mock: HTTPXMock) -> None:
     httpx_mock.add_response(status_code=500, url=ALGOD_HEALTH_URL)
     result = invoke("localnet start")
@@ -98,7 +126,7 @@ def test_localnet_start_health_bad_status(app_dir_mock: AppDirs, httpx_mock: HTT
     )
 
 
-@pytest.mark.usefixtures("_localnet_up_to_date")
+@pytest.mark.usefixtures("_localnet_up_to_date", "_mock_proc_with_running_localnet")
 def test_localnet_start_failure(app_dir_mock: AppDirs, proc_mock: ProcMock) -> None:
     proc_mock.should_bad_exit_on("docker compose up")
 
@@ -108,7 +136,7 @@ def test_localnet_start_failure(app_dir_mock: AppDirs, proc_mock: ProcMock) -> N
     verify(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}").replace("\\", "/"))
 
 
-@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_up_to_date")
+@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_up_to_date", "_mock_proc_with_running_localnet")
 def test_localnet_start_up_to_date_definition(app_dir_mock: AppDirs) -> None:
     (app_dir_mock.app_config_dir / "sandbox").mkdir()
     (app_dir_mock.app_config_dir / "sandbox" / "docker-compose.yml").write_text(get_docker_compose_yml())
@@ -121,7 +149,7 @@ def test_localnet_start_up_to_date_definition(app_dir_mock: AppDirs) -> None:
     verify(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}").replace("\\", "/"))
 
 
-@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_up_to_date")
+@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_up_to_date", "_mock_proc_with_running_localnet")
 def test_localnet_start_out_of_date_definition(app_dir_mock: AppDirs) -> None:
     (app_dir_mock.app_config_dir / "sandbox").mkdir()
     (app_dir_mock.app_config_dir / "sandbox" / "docker-compose.yml").write_text("out of date config")
@@ -143,7 +171,7 @@ def test_localnet_start_out_of_date_definition(app_dir_mock: AppDirs) -> None:
     )
 
 
-@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_up_to_date")
+@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_up_to_date", "_mock_proc_with_running_localnet")
 def test_localnet_start_out_of_date_definition_and_missing_config(app_dir_mock: AppDirs) -> None:
     (app_dir_mock.app_config_dir / "sandbox").mkdir()
     (app_dir_mock.app_config_dir / "sandbox" / "docker-compose.yml").write_text("out of date config")
@@ -162,7 +190,7 @@ def test_localnet_start_out_of_date_definition_and_missing_config(app_dir_mock: 
     )
 
 
-@pytest.mark.usefixtures("app_dir_mock")
+@pytest.mark.usefixtures("app_dir_mock", "_mock_proc_with_running_localnet")
 def test_localnet_start_without_docker(proc_mock: ProcMock) -> None:
     proc_mock.should_fail_on("docker compose version")
 
@@ -172,7 +200,7 @@ def test_localnet_start_without_docker(proc_mock: ProcMock) -> None:
     verify(result.output)
 
 
-@pytest.mark.usefixtures("app_dir_mock")
+@pytest.mark.usefixtures("app_dir_mock", "_mock_proc_with_running_localnet")
 def test_localnet_start_without_docker_compose(proc_mock: ProcMock) -> None:
     proc_mock.should_bad_exit_on("docker compose version")
 
@@ -182,7 +210,7 @@ def test_localnet_start_without_docker_compose(proc_mock: ProcMock) -> None:
     verify(result.output)
 
 
-@pytest.mark.usefixtures("app_dir_mock")
+@pytest.mark.usefixtures("app_dir_mock", "_mock_proc_with_running_localnet")
 def test_localnet_start_without_docker_engine_running(proc_mock: ProcMock) -> None:
     proc_mock.should_bad_exit_on("docker version")
 
@@ -192,7 +220,7 @@ def test_localnet_start_without_docker_engine_running(proc_mock: ProcMock) -> No
     verify(result.output)
 
 
-@pytest.mark.usefixtures("app_dir_mock")
+@pytest.mark.usefixtures("app_dir_mock", "_mock_proc_with_running_localnet")
 def test_localnet_start_with_old_docker_compose_version(proc_mock: ProcMock) -> None:
     proc_mock.set_output("docker compose version --format json", [json.dumps({"version": "v2.2.1"})])
 
@@ -202,7 +230,7 @@ def test_localnet_start_with_old_docker_compose_version(proc_mock: ProcMock) -> 
     verify(result.output)
 
 
-@pytest.mark.usefixtures("_health_success", "_localnet_up_to_date")
+@pytest.mark.usefixtures("_health_success", "_localnet_up_to_date", "_mock_proc_with_running_localnet")
 def test_localnet_start_with_unparseable_docker_compose_version(app_dir_mock: AppDirs, proc_mock: ProcMock) -> None:
     proc_mock.set_output("docker compose version --format json", [json.dumps({"version": "v2.5-dev123"})])
 
@@ -212,7 +240,7 @@ def test_localnet_start_with_unparseable_docker_compose_version(app_dir_mock: Ap
     verify(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}").replace("\\", "/"))
 
 
-@pytest.mark.usefixtures("_health_success", "_localnet_up_to_date")
+@pytest.mark.usefixtures("_health_success", "_localnet_up_to_date", "_mock_proc_with_running_localnet")
 def test_localnet_start_with_gitpod_docker_compose_version(app_dir_mock: AppDirs, proc_mock: ProcMock) -> None:
     proc_mock.set_output("docker compose version --format json", [json.dumps({"version": "v2.10.0-gitpod.0"})])
 
@@ -222,7 +250,7 @@ def test_localnet_start_with_gitpod_docker_compose_version(app_dir_mock: AppDirs
     verify(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}").replace("\\", "/"))
 
 
-@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_out_of_date")
+@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_out_of_date", "_mock_proc_with_running_localnet")
 def test_localnet_start_out_date(app_dir_mock: AppDirs) -> None:
     result = invoke("localnet start")
 
@@ -230,7 +258,9 @@ def test_localnet_start_out_date(app_dir_mock: AppDirs) -> None:
     verify(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}").replace("\\", "/"))
 
 
-@pytest.mark.usefixtures("proc_mock", "_health_success", "_localnet_img_check_cmd_error")
+@pytest.mark.usefixtures(
+    "proc_mock", "_health_success", "_localnet_img_check_cmd_error", "_mock_proc_with_running_localnet"
+)
 def test_localnet_img_check_cmd_error(app_dir_mock: AppDirs) -> None:
     result = invoke("localnet start")