+++ tests/localnet/test_localnet_stop.py
@@ -1,3 +1,5 @@
+import json
+
 import pytest
 
 from tests.utils.app_dir_mock import AppDirs
@@ -6,7 +8,7 @@ from tests.utils.click_invoker import invoke
 from tests.utils.proc_mock import ProcMock
 
 
-@pytest.mark.usefixtures("proc_mock")
+@pytest.mark.usefixtures("proc_mock", "_mock_proc_with_running_localnet")
 def test_localnet_stop(app_dir_mock: AppDirs) -> None:
     (app_dir_mock.app_config_dir / "sandbox").mkdir()
     (app_dir_mock.app_config_dir / "sandbox" / "docker-compose.yml").write_text("existing")
@@ -18,6 +20,31 @@ def test_localnet_stop(app_dir_mock: AppDirs) -> None:
     verify(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}").replace("\\", "/"))
 
 
+def test_localnet_stop_with_name(app_dir_mock: AppDirs, proc_mock: ProcMock) -> None:
+    (app_dir_mock.app_config_dir / "sandbox_test").mkdir()
+    (app_dir_mock.app_config_dir / "sandbox_test" / "docker-compose.yml").write_text("existing")
+    (app_dir_mock.app_config_dir / "sandbox_test" / "algod_config.json").write_text("existing")
+    proc_mock.set_output(
+        "docker compose ls --format json --filter name=algokit_sandbox*",
+        [
+            json.dumps(
+                [
+                    {
+                        "Name": "algokit_sandbox_test",
+                        "Status": "running",
+                        "ConfigFiles": "sandbox_test/docker-compose.yml",
+                    }
+                ]
+            )
+        ],
+    )
+    result = invoke("localnet stop")
+
+    assert result.exit_code == 0
+    verify(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}").replace("\\", "/"))
+
+
+@pytest.mark.usefixtures("_mock_proc_with_running_localnet")
 def test_localnet_stop_failure(app_dir_mock: AppDirs, proc_mock: ProcMock) -> None:
     (app_dir_mock.app_config_dir / "sandbox").mkdir()
     (app_dir_mock.app_config_dir / "sandbox" / "docker-compose.yml").write_text("existing")
@@ -30,7 +57,7 @@ def test_localnet_stop_failure(app_dir_mock: AppDirs, proc_mock: ProcMock) -> No
     verify(result.output.replace(str(app_dir_mock.app_config_dir), "{app_config}").replace("\\", "/"))
 
 
-@pytest.mark.usefixtures("proc_mock")
+@pytest.mark.usefixtures("proc_mock", "_mock_proc_with_running_localnet")
 def test_localnet_stop_no_existing_definition(app_dir_mock: AppDirs) -> None:
     result = invoke("localnet stop")